module, loader: abstract away nickname_commands
Having the @module.nickname_commands() decorator build its input
directly into a rule made getting at the internals of how it was built
pretty much impossible. Giving callables a 'nickname_commands' attribute
to go with the 'commands' attribute they already have makes sense, and
lets the rule-building happen in the same place (the loader) for both
types of command.

Moving all of the code that builds regexes for rules into sopel.tools
also lets tests import those functions and use them directly, instead of
trying to hack together a test that somehow kludges together functions
from the loader, module decorators, etc. This is more sensible than
having one command regex type (`.command`s) defined in sopel.tools, and
the other (`Sopel: command`s) hard-coded in the module decorator.
