+++ sopel/loader.py
@@ -6,13 +6,10 @@ import os.path
 import re
 import sys
 
-from sopel.tools import itervalues, get_command_regexp
+from sopel.tools import compile_rule, itervalues, get_command_regexp, get_nickname_command_regexp
 
 if sys.version_info.major >= 3:
     basestring = (str, bytes)
-
-# Can be implementation-dependent
-_regex_type = type(re.compile(''))
 
 
 def get_module_description(path):
@@ -110,27 +107,6 @@ def enumerate_modules(config, show_all=False):
     return modules
 
 
-def compile_rule(nick, pattern, alias_nicks):
-    # Not sure why this happens on reloads, but it shouldn't cause problems…
-    if isinstance(pattern, _regex_type):
-        return pattern
-
-    if alias_nicks:
-        nicks = alias_nicks.copy()
-        nicks.append(nick)
-        nicks = map(re.escape, nicks)
-        nick = '(?:%s)' % '|'.join(nicks)
-    else:
-        nick = re.escape(nick)
-
-    pattern = pattern.replace('$nickname', nick)
-    pattern = pattern.replace('$nick', r'{}[,:]\s+'.format(nick))
-    flags = re.IGNORECASE
-    if '\n' in pattern:
-        flags |= re.VERBOSE
-    return re.compile(pattern, flags)
-
-
 def trim_docstring(doc):
     """Get the docstring as a series of lines that can be sent"""
     if not doc:
@@ -183,10 +159,13 @@ def clean_callable(func, config):
             func.rule = [func.rule]
         func.rule = [compile_rule(nick, rule, alias_nicks) for rule in func.rule]
 
-    if hasattr(func, 'commands'):
+    if hasattr(func, 'commands') or hasattr(func, 'nickname_commands'):
         func.rule = getattr(func, 'rule', [])
-        for command in func.commands:
+        for command in getattr(func, 'commands', []):
             regexp = get_command_regexp(prefix, command)
+            func.rule.append(regexp)
+        for command in getattr(func, 'nickname_commands', []):
+            regexp = get_nickname_command_regexp(nick, command, alias_nicks)
             func.rule.append(regexp)
         if hasattr(func, 'example'):
             example = func.example[0]["example"]
@@ -194,7 +173,10 @@ def clean_callable(func, config):
             if example[0] != help_prefix and not example.startswith(nick):
                 example = help_prefix + example[len(help_prefix):]
         if doc or example:
-            for command in func.commands:
+            cmds = []
+            cmds.extend(getattr(func, 'commands', []))
+            cmds.extend(getattr(func, 'nickname_commands', []))
+            for command in cmds:
                 func._docs[command] = (doc, example)
 
     if hasattr(func, 'intents'):
@@ -214,7 +196,7 @@ def load_module(name, path, type_):
 
 
 def is_triggerable(obj):
-    return any(hasattr(obj, attr) for attr in ('rule', 'intents', 'commands'))
+    return any(hasattr(obj, attr) for attr in ('rule', 'intents', 'commands', 'nickname_commands'))
 
 
 def clean_module(module, config):
